requirements.txt
streamlit
pandas
matplotlib
seaborn
google-generativeai
sentence-transformers
faiss-cpu
rank-bm25
numpy